Added in convenience functions to speed up the generation of large numbers of random numbers, and have also fixed a small bug in MultiDouble that was casting an array of doubles to floats, before then storing them as doubles...
Skipping travis until I commit in the python wrappers

[ci skip]

+++ corelib/src/libs/SireMaths/multidouble.cpp
@@ -146,7 +146,7 @@ MultiDouble::MultiDouble(const double *array, int size)
     }
     else
     {
-        float tmp[MULTIFLOAT_SIZE];
+        double tmp[MULTIFLOAT_SIZE];
         
         for (int i=0; i<size; ++i)
         {
